Added spec using partial sums.
Partial sums with the help of quantifiers are powerful enough
to specifiy in the invariant that something is really the sum of
all elements in an array.  With the help of monotonicity (if all
summed elements are nonnegative) we can even show that the sum is
larger than the sum of some individuals.